+++ fmt/fmt_strn.c
@@ -10,5 +10,5 @@ size_t fmt_strn(char *out,const char *in,size_t limit) {
     if (!*t || t==u) break; if (s) { *s=*t; ++s; } ++t;
     if (!*t || t==u) break; if (s) { *s=*t; ++s; } ++t;
   }
-  return t-in;
+  return (size_t)(t-in);
 }
